requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
djangorestframework-simplejwt
Pillow
channels
channels-redis
drf-spectacular
daphne
drf-yasg
celery
redis
gunicorn==22.0.0
django_extensions
dj-database-url
django-filter
